Separate style code to separate file with name style.js and add dev element with some css style Relates #57
